extract headingTwo in to a component
